a number of issues:
        - Issue 44 ("Break tags are stripped"): Mozilla browsers keep 
around a BR tag at the end of paragraphs. When a user would insert a 
line break at the end of one, it resulted in two BR tags in a row, which 
the block cleanup code in Util.Block treated as a paragraph separator, 
and it would remove both the Mozilla BR tag and the one that the user 
created. To work around this issue, the block cleanup code now ignores 
two consecutive BR's that are at the end of a paragraph.

	- Issue 45 ("Paragraphs wrapped around pasted text even 
when copied text is plain string"): Copied and pasted HTML is cleaned up 
by first inserting it into a DIV tag on the editing document and then 
running UI.Clean.clean on it. Because the block cleanup rules treat 
DIV's as a "paragraph container" (which means that all inline content 
directly within it must be instead wrapped in paragraphs), if a user 
copied only some inline content (e.g. text) it would be wrapped into a 
paragraph. Similarly, pasted inline content would also be wrapped.
	To work around this, Util.Block.enforce_rules (and 
UI.Clean.clean) now accept a list of overrides that can change how it 
behaves towards certain tags, and the clipboard cleanup code now 
specifies that DIV tags are normal blocks.
		
	- Issue 46 ("Pasted content improperly ordered"): This 
was due to a simple bug in the paragraph rule enforcer.
		
	- Issue 47 ("HTML cleanup button should only run 
appropriate actions"): UI.Clean.clean previously accepted an "automatic" 
parameter, which specified whether or not the cleanup was being run 
automatically. It has now been renamed to "live" (whether or not the 
cleanup is being run on content that is actively being edited), which 
better reflects why the distinction exists. The cleanup button now 
specifies that it is being run on live content.